Create 1. Gate Level Modeling - 4x1 MUX
MUX- It is a combinational circuit which have many data inputs and single output depending on control or select inputs.
For N input lines, log n (base2) selection lines, or we can say that for 2n input lines, n selection lines are required.

Here, a 4x1 MUX is designed using 2 NOT gates, 4 three-input AND gates and 1 OR gate. 
i0,i1,i2,i3 are connected to the four AND gates respectively. s0,s1 are the select lines connected to the NOT gates, s0n and s1n are the outputs of the NOT gate. 

1st AND gate has inputs- i0,s1n,s0n and output y0
2nd AND gate has inputs- i1,s1n,s0 and output y1
3rd AND gate has inputs - i2,s1,s0n and output y2
4th AND gate has inputs - i3,s1,s0 and output y3

y0,y1,y2,y3 are connected to OR gate whose output is called 'out'.

Out = i0s1ns0n + i1s1ns0 + i2s1ns0 + i3s1s0